systemtest python-bareos: split unittests file
Split monolitic systemtests/tests/python-bareos/python-bareos-unittest.py
into one file per test class.
Move commonly used functionality to modules in
systemtests/python-modules/bareos_unittest/.
Use ``unittest discover`` to run all Python tests in the directory named test_*.py.
