--- systemtests/tests/python-bareos/testrunner
@@ -1,47 +1,0 @@
-#!/bin/bash
-set -e
-set -o pipefail
-set -u
-#
-TestName="$(basename "$(pwd)")"
-export TestName
-
-JobName=backup-bareos-fd
-#shellcheck source=../environment.in
-. ./environment
-
-JobName=backup-bareos-fd
-#shellcheck source=../scripts/functions
-. "${rscripts}"/functions
-"${rscripts}"/cleanup
-"${rscripts}"/setup
-
-# Fill ${BackupDirectory} with data.
-setup_data
-
-start_test
-
-start_bareos
-
-# allow daemons to start listening before python connects
-tries=60
-echo "waiting for director to start"
-while ! bin/bconsole <<< "quit" > /dev/null 2>&1; do
-  [ $((tries-=1)) -eq 0 ] && {
-    echo "could not connect to Director"
-    exit 1
-  }
-  echo -n "."
-  sleep 1
-done
-
-
-
-
-run_python_unittests "$@"
-
-check_for_zombie_jobs storage=File client=bareos-fd
-
-stop_bareos
-
-end_test
