+++ systemtests/tests/python-bareos/testrunner-unittest
@@ -1,0 +1,19 @@
+#!/bin/bash
+set -e
+set -o pipefail
+set -u
+
+#shellcheck source=../environment.in
+. ./environment
+#shellcheck source=../scripts/functions
+. "${rscripts}"/functions
+
+TestName="$(get_test_name "$0")"
+
+start_test
+
+run_python_unittests "$@"
+
+check_for_zombie_jobs storage=File client=bareos-fd
+
+end_test
